+++ tests/misc.py
@@ -1,3 +1,4 @@
+import pickle
 import typing
 import unittest
 
@@ -752,3 +753,42 @@ class MiscTests(unittest.TestCase):
         _, samples = sample(mcs, rngs, n_samples=20000, thinning=10)
         actual_rhat = rhat(samples)
         self.assertTrue(np.all(np.isclose(actual_rhat, expected_rhat)))
+
+    def test_pickling_of_PyParallelTemperingChain(self):
+        replicates = 3
+        n_temps = 4
+
+        problem = Problem([[1, 1]], [1], Gaussian(), starting_point=[0, 0])
+
+        sync_rngs = [RandomNumberGenerator(seed=1991 + r) for r in range(replicates)]
+
+        temperature_ladder = [1.0 - float(n) / (n_temps - 1) for n in range(n_temps)]
+
+        mcs = [
+            MarkovChain(
+                proposal=UniformCoordinateHitAndRunProposal,
+                problem=problem,
+                starting_point=np.zeros(2),
+            )
+            for r in range(replicates)
+        ]
+
+        # Creates one parallel tempering ensemble for each replicate.
+        # Each ensemble will have len(temperature_ladder) chains.
+        chains = create_py_parallel_tempering_ensembles(
+            markov_chains=mcs,
+            temperature_ladder=temperature_ladder,
+            sync_rngs=sync_rngs,
+            draws_per_exchange_attempt=100,
+        )
+
+        dump = pickle.dumps(chains)
+        del chains
+        new_chains = pickle.loads(dump)
+
+        # self.assertTrue(len(new_chains) == len(chains))
+        # for i in range(len(new_chains)):
+        #     self.assertTrue(np.all(np.isclose(chains[i].state, new_chains[i].state)))
+        #     print('index', new_chains[i].chain_index)
+        #     print('coldness', new_chains[i].coldness)
+        #     print('sm', new_chains[i].shared_memory)