+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Latest
+- [#158](https://jugit.fz-juelich.de/IBG-1/ModSim/hopsy/-/issues/158): Fix pickling of parallel tempered chains
+
 ## v1.5.1
 - [#138](https://jugit.fz-juelich.de/IBG-1/ModSim/hopsy/-/issues/138): Ignore fixed dimensions when computing ESS & rhat
 - [#157](https://jugit.fz-juelich.de/IBG-1/ModSim/hopsy/-/issues/157): Pin numpy<2.0